+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+
+## [22.1.0][22.1.0] - 2016-11-29
 ### Added
 - [Feature] Added `min` and `max` props to `Input` to be used when `type` is `number`. (#676)
 
@@ -703,7 +705,7 @@ _This version was accidentally released and contains no changes._
 ### Added
 - [Patch] Use semantic versioning in LEGO (#58).
 
-[Unreleased]: https://github.com/optimizely/oui/compare/v22.0.0...HEAD
+[Unreleased]: https://github.com/optimizely/oui/compare/v22.1.0...HEAD
 [0.0.2]: https://github.com/optimizely/oui/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/optimizely/oui/compare/v0.0.2...v0.0.3
 [1.0.0]: https://github.com/optimizely/oui/compare/v0.0.3...v1.0.0
@@ -774,3 +776,4 @@ _This version was accidentally released and contains no changes._
 [20.0.0]: https://github.com/optimizely/oui/compare/v19.3.0...v20.0.0
 [21.0.0]: https://github.com/optimizely/oui/compare/v20.0.0...v21.0.0
 [22.0.0]: https://github.com/optimizely/oui/compare/v21.0.0...v22.0.0
+[22.1.0]: https://github.com/optimizely/oui/compare/v22.0.0...v22.1.0